+++ Project/source/Managers/EnemyManager.h
@@ -26,7 +26,7 @@ public:
 
 	void Spawn();
 
-	int GetEnemyNum() { return m_pEnemies.size(); }
+	int GetEnemyNum() { return static_cast<int>(m_pEnemies.size()); }
 
 private:
 	Vector3 GetSpawnPos();
